Migrate hw-reset/test-t2enum.py to pytest
